+++ web-frontend/modules/core/components/group/GroupContext.vue
@@ -2,25 +2,31 @@
   <Context ref="context">
     <div class="context__menu-title">{{ group.name }} ({{ group.id }})</div>
     <ul class="context__menu">
-      <li>
+      <li v-if="group.permissions === 'ADMIN'">
         <a @click="$emit('rename')">
           <i class="context__menu-icon fas fa-fw fa-pen"></i>
           Rename group
         </a>
       </li>
-      <li>
+      <li v-if="group.permissions === 'ADMIN'">
         <a @click=";[$refs.groupMembersModal.show(), hide()]">
           <i class="context__menu-icon fas fa-fw fa-users"></i>
           Members
         </a>
       </li>
-      <li>
+      <li v-if="group.permissions === 'ADMIN'">
         <a @click="showGroupTrashModal">
           <i class="context__menu-icon fas fa-fw fa-recycle"></i>
           View trash
         </a>
       </li>
       <li>
+        <a @click="$refs.leaveGroupModal.show()">
+          <i class="context__menu-icon fas fa-fw fa-door-open"></i>
+          Leave group
+        </a>
+      </li>
+      <li v-if="group.permissions === 'ADMIN'">
         <a
           :class="{ 'context__menu-item--loading': loading }"
           @click="deleteGroup"
@@ -31,10 +37,17 @@
       </li>
     </ul>
     <GroupMembersModal
+      v-if="group.permissions === 'ADMIN'"
       ref="groupMembersModal"
       :group="group"
     ></GroupMembersModal>
-    <TrashModal ref="groupTrashModal" :initial-group="group"> </TrashModal>
+    <TrashModal
+      v-if="group.permissions === 'ADMIN'"
+      ref="groupTrashModal"
+      :initial-group="group"
+    >
+    </TrashModal>
+    <LeaveGroupModal ref="leaveGroupModal" :group="group"></LeaveGroupModal>
   </Context>
 </template>
 
@@ -43,10 +56,11 @@ import GroupMembersModal from '@baserow/modules/core/components/group/GroupMembe
 import context from '@baserow/modules/core/mixins/context'
 import { notifyIf } from '@baserow/modules/core/utils/error'
 import TrashModal from '@baserow/modules/core/components/trash/TrashModal'
+import LeaveGroupModal from '@baserow/modules/core/components/group/LeaveGroupModal'
 
 export default {
   name: 'GroupContext',
-  components: { TrashModal, GroupMembersModal },
+  components: { LeaveGroupModal, TrashModal, GroupMembersModal },
   mixins: [context],
   props: {
     group: {