+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Made it possible to leave a group.
 * Changed web-frontend `/api/docs` route into `/api-docs`.
 * Bumped the dependencies.
 * The internal setting allowing Baserow to run with the user tables in a separate